chore: move allow list endpoint
